pyproject.toml
[tool.poetry]
name = "librispeech"
version = "0.1.0"
description = ""
authors = ["Astellon <keitagoto@fastmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.11.0"
torchaudio = "^0.11.0"
PySoundFile = "^0.9.0"
tqdm = "^4.64.0"
webdataset = "^0.2.5"
audioread = "^2.1.9"
pytorch-lightning = "^1.6.3"
scikit-learn = "^1.0.2"
torchmetrics = "^0.8.2"
wandb = "^0.12.16"
hydra-core = "^1.1.2"

[tool.poetry.dev-dependencies]
docker = "^5.0.3"
google-cloud-compute = "^1.3.0"
google-cloud-artifact-registry = "^1.2.0"
torchinfo = "^1.6.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
